Move package config/file -> internal/configfile
