+++ internal/configfile/find.go
@@ -1,4 +1,4 @@
-package file
+package configfile
 
 import "os"
