+++ internal/configfile/testdata/extends/extends-circular-1.yml
