+++ internal/configfile/testdata/extends/extends-circular-2.yml
