+++ internal/configfile/testdata/extends/extends-circular-self.yml
