+++ internal/configfile/testdata/invalid/badfields.json
