+++ internal/configfile/testdata/valid/benchttp.json
