+++ internal/configfile/testdata/valid/benchttp.yaml
